+++ src/components/HistoricalPriceComponent.js
