+++ src/pages/TransactionPage.js
@@ -1,14 +1,14 @@
 import React from "react";
 import { withRouter } from "react-router-dom";
-import Priceshow from "../components/PriceshowComponent";
+import PriceShow from "../components/PriceShowComponent";
 import BuySell from "../components/BuySellComponent"
 
 class TransactionPage extends React.Component {
     render() {
         return (
             <div className = "MainContainer">
-                <div className = "PriceshowContainer">
-                    <Priceshow />
+                <div className = "PriceShowContainer">
+                    <PriceShow />
                 </div>
                 <br></br>
                 <br></br>
